procfs_read_task_info() and procfs_read_thread_info(): flags should be of type uint64_t, not int

+++ kext/procfs_data.c
@@ -225,7 +225,7 @@ procfs_read_task_info(procfsnode_t *pnp, uio_t uio, __unused vfs_context_t ctx)
     int error = 0;
     proc_t p = proc_find(pnp->node_id.nodeid_pid);
     int pid = proc_pid(p);
-    int flags = 0;
+    uint64_t flags = 0;
 
     if (p != NULL) {
         struct proc_taskinfo info;
@@ -253,7 +253,7 @@ procfs_read_thread_info(procfsnode_t *pnp, uio_t uio, __unused vfs_context_t ctx
     int error = 0;
     proc_t p = proc_find(pnp->node_id.nodeid_pid);
     int pid = proc_pid(p);
-    int flags = 0;
+    uint64_t flags = 0;
 
     if (p != NULL) {
         struct proc_threadinfo info;
